Initial skeleton of `ants_ai` the Python package for interacting w/ the game
This skeleton follows the structure required by Pyo3 so that we can
generate Python bindings for the `ants_engine` written in Rust, and
use them in the `ants_ai` Python package to implement an RL environment,
etc.

I wanted to keep `ants_engine` as its own folder but it wasn't playing
nicely with Pyo3 so this works much better.