[WIP] Change observation to `dict` to provide a mask of live ants
This mask will be used during learning and inferencing to generate
actions only for live ("valid") ants.

This commit also adds a base `Agent` class and implements a
`RandomAgent` class that shows how this mask is used during
inference. `ants_ai` will provide a `RandomAgent` class that users
can leverage as the agents for other players.